pyproject.toml
[tool.poetry]
name = "url-to-calibre"
version = "1.0.0"
description = "Convert a URL's contents into ebook format and add to Calibre library"
authors = ["Alex C-G"]
readme = "README.md"
packages = [{ include = "url_to_calibre" }]

[tool.poetry.dependencies]
python = ">=3.12"
newspaper3k = "^0.2.8"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
ebooklib = "^0.18"
lxml-html-clean = "^0.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
url-to-calibre = "url_to_calibre:main"
